% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contest.R
\name{ord}
\alias{ord}
\alias{dord}
\alias{pord}
\title{Order statistics density and distribution functions}
\usage{
dord(x, k, n, p, d, ...)

pord(x, k, n, p, ...)
}
\arguments{
\item{x}{vector of values}

\item{k, n}{order statistic and number of iid random variables}

\item{p, d}{distribution and density}

\item{...}{further arguments for distribution and density}
}
\value{
Probability of the k-th order statistic from n iid random variables
}
\description{
Density and distribution function of the  `k`'s order statistic for `n` iid random variables drawn from a given distribution `p` with density `d`.
}
\examples{
curve(dord(x, 1, 10, pnorm, dnorm, mean=0, sd=1), from=-4, to=4)
curve(dord(x, 10, 10, pnorm, dnorm, mean=0, sd=1), add=TRUE, lty=2)
curve(dnorm(x, mean=0, sd=1), add=TRUE, lty=3)
title("Density order statistics from 10 normal r.v.")
legend("topright", c("Min (1st OS)", "Max (n-th OS)", "Normal density"), lty=1:3, bty="n")
}
\keyword{order}
\keyword{statistics}
